Fix type mismatches

+++ src/appleseedmaya/exporters/mandelbrotexporter.cpp
@@ -127,7 +127,7 @@ void MandelbrotExporter::exportParameterValue(
         std::vector<MandelbrotColorsEntry> mandelbrotColors;
         mandelbrotColors.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -175,7 +175,7 @@ void MandelbrotExporter::exportParameterValue(
         std::vector<MandelbrotValuesEntry> mandelbrotValues;
         mandelbrotValues.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
