+++ src/appleseedmaya/exporters/remaphsvexporter.cpp
@@ -108,7 +108,7 @@ void RemapHsvExporter::exportParameterValue(
         std::vector<RemapHsvEntry> remapHue;
         remapHue.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -154,7 +154,7 @@ void RemapHsvExporter::exportParameterValue(
         std::vector<RemapHsvEntry> remapSaturation;
         remapSaturation.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -200,7 +200,7 @@ void RemapHsvExporter::exportParameterValue(
         std::vector<RemapHsvEntry> remapValue;
         remapValue.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
